==> 99_check_wayback.ipynb <==
import waybackpy as wbp
import savepagenow as spn
import pandas as pd
import mariadb
import getpass
import mysql.connector

pw = getpass.getpass()
con = mysql.connector.connect(
# mariadb.connect(
    user = "scraper",
    password = pw,
    host = "127.0.0.1",
    port = 3336,
    database = "russian_media_database_rt"
)
cursor = con.cursor()

# get domains:
versions = pd.read_sql_query("SELECT * FROM sitemap_versions", con)
versions = versions.link.str.slice(0,-12)

done = pd.read_sql_query("SELECT link FROM archived_pages", con)
len(done)

# get all archived urls: 

from sqlalchemy import create_engine


username = "scraper"
host = "127.0.0.1"
port = "3336"
database = "russian_media_database_rt"

# Create a connection string
connection_str = f"mysql+mysqlconnector://{username}:{pw}@{host}:{port}/{database}"

# Create an SQLAlchemy engine
engine = create_engine(connection_str)



versions = versions[4]

versions = [versions, "https://lat.rt.rs"]
versions


for base_url in versions:

    with engine.begin() as connection:

        archive_urls = []

        wayback_base = wbp.Url(base_url)
        archive = wayback_base.known_urls(subdomain=True)

        for url in archive:
            # print(url)
            # write to database:
            archive_urls.append(url)

        archive_urls = [url for url in archive_urls if url not in done]

        archive_urls_df = pd.DataFrame(archive_urls, columns=['link'])

        archive_urls_df.to_sql('archived_pages', con = connection, if_exists = 'append', index=False , chunksize = 1000)

        connection.commit()
    #    for url in archive_urls:
    #        cursor.execute("INSERT INTO archived_pages (link) VALUES (%s)", (url,))
    #        con.commit()    

        print("saved",  base_url, len(archive_urls))Milan Schröder

archive_urls = pd.read_sql_query("SELECT * FROM archived_pages;", con)
len(archive_urls)

# create df:
url_df = pd.DataFrame({'link' = archive_urls, 'archived': 1})  

# compare to self-scraped pages: 

url_list = pd.read_sql_query("SELECT link, doc_hash, available_online FROM page_data;", con)
url_list

# join archived and scraped pages:
# how many of the scarped pages are archived?
# how many of the archived pages are scraped?

compare = pd.read_sql_query("""
    SELECT * link, available_online, doc_hash
    FROM page_data
    FULL JOIN archived_pages
    ON page_data.link = archived_pages.link
    ;
""", con)

# write to database:
for i in range(len(archive_urls)):
    cursor.execute("INSERT INTO archived_pages (link) VALUES ('{archive_urls[i]}')")

# loop over df:
for index, row in url_df.iterrows():
    cursor.execute("INSERT INTO archived_pages (link, archived) VALUES (?, ?)", row.link, row.archived)

# create table
sql_command = """CREATE TABLE IF NOT EXISTS archived_pages (
	rowid BIGINT(20) NOT NULL AUTO_INCREMENT,
    link TEXT,
	PRIMARY KEY (`rowid`) USING BTREE
    );
"""

# execute the statement
cursor.execute(sql_command)

# commit changes:
con.commit()


# clsoe connection:
con.close()
