==> 99_save_archive.ipynb <==
import savepagenow as spn
import pandas as pd
from getpass import getpass
import mariadb

!pip install savepagenow

pipenv install savepagenow

Already archived urls:

import waybackpy

base_url = "https://eccfp.eu/"

wayback.waybackpy.Url(url = base_url)
known_urls = wayback.known_urls(subdomain = True)
